requirements.txt
click
fastapi
greenlet
h11
pydantic
PyMySQL
PyYAML
SQLAlchemy
starlette
typing-extensions
uvicorn
requests
pysftp
mysql-connector-python
gunicorn
python-dotenv
json2html